unify factories to a single module
